pyproject.toml
[tool.poetry]
name = "repository-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Dmitrii Zakharov <dmitrii@zakharov.cc>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.7.4.post0"
aioredis = "1.3.1"
Jinja2 = "3.0.1"
first = "2.0.2"

[tool.poetry.dev-dependencies]
pytest-aiohttp = "0.3.0"
mypy = "0.812"
black = "21.5b1"
flake8 = "3.9.2"
isort = "5.8.0"
pre-commit = "2.13.0"
bandit = "1.7.0"
safety = "1.10.3"
pydocstyle = "6.1.1"
pytest-cov = "2.12.0"

[tool.black]
skip-string-normalization = 1

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
